+++ core/src/components/accordion/test/a11y/accordion.e2e.ts
@@ -3,10 +3,11 @@ import { configs, test } from '@utils/test/playwright';
 
 configs().forEach(({ config, title }) => {
   test.describe(title('accordion: a11y'), () => {
-    test('accordions should be keyboard navigable', async ({ page, skip, browserName }) => {
-      // TODO(FW-1764): remove skip once issue is resolved
+    // TODO(ROU-8157): remove skip once the keyboard navigation is working again
+    test.skip('accordions should be keyboard navigable', async ({ page, skip, browserName }) => {
+      // TODO(ROU-5358): remove skip once issue is resolved
       skip.browser('firefox', 'https://github.com/ionic-team/ionic-framework/issues/25070');
-      // TODO (FW-2979)
+      // TODO (ROU-5437)
       skip.browser('webkit', 'Safari 16 only allows text fields and pop-up menus to be focused.');
 
       await page.goto(`/src/components/accordion/test/a11y`, config);