+++ packages/angular/test/apps/ng18/e2e/src/lazy/modal-nav-params.spec.ts
@@ -1,0 +1,20 @@
+describe('Modal Nav Params', () => {
+
+  beforeEach(() => {
+    cy.visit('/lazy/version-test/modal-nav-params');
+  });
+
+  it('should assign the rootParams when presented in a modal multiple times', () => {
+    cy.contains('Open Modal').click();
+    cy.get('ion-modal').should('exist').should('be.visible');
+    cy.get('ion-modal').contains('OK');
+
+    cy.contains("Close").click();
+    cy.get('ion-modal').should('not.be.visible');
+
+    cy.contains('Open Modal').click();
+    cy.get('ion-modal').should('exist').should('be.visible');
+    cy.get('ion-modal').contains('OK').should('exist');
+  });
+
+});